requirements.txt
pymodbus>=3.6.0
websocket-client>=1.6.0
pycryptodome>=3.20.0
aiohttp>=3.8.5